+++ src/Components/History/HistoryCard.jsx
@@ -1,11 +1,11 @@
 import { Link } from "react-router-dom";
 import { useDataContext } from "../../Context/data-context";
-import { AddToPlaylist } from "../Playlist/addToPlaylist";
-import { imageURL } from "../Video/videoUtil";
+import { AddToPlaylist } from "../Playlist/AddToPlaylist";
+import { imageURL, videoExists } from "../Video/videoUtil";
 
-export const LikedVideoCard = ({ vid }) => {
+export const HistoryCard = ({ vid }) => {
   const {
-    state: { videoList },
+    state: { videoList, likedVideos },
     dispatch,
   } = useDataContext();
 
@@ -28,15 +28,24 @@ export const LikedVideoCard = ({ vid }) => {
       <small className="txt-small txt-grey">{author}</small>
       </Link>
       <small className="txt-grey">
+        <i
+          onClick={() => dispatch({ type: "TOGGLE_LIKE", payload: vid })}
+          className={
+            videoExists(likedVideos, vid)
+              ? "fas fa-thumbs-up primaryBg-txt"
+              : "fas fa-thumbs-up"
+          }
+        ></i>
       <AddToPlaylist vid={vid}/>
       </small>
       <button type="button" onClick={() => dispatch({
-                        type: "TOGGLE_LIKE",
+                        type: "REMOVE_FROM_HISTORY",
                         payload: vid 
                       })}
             className="btn btn-secondary btn-dismiss"><i className="fas fa-times"></i></button>
   
     </div>
+
   );
 };
