+++ src/Components/LikedVideos/LikedVideos.jsx
@@ -1,7 +1,7 @@
 import { useEffect } from "react";
 import { Link } from "react-router-dom";
 import { useDataContext } from "../../Context/data-context"
-import { LikedVideoCard } from "./likedVideoCard";
+import { LikedVideoCard } from "./LikedVideoCard";
 import like from "../../images/like.svg";
 
 export const LikedVideos = () => {
