+++ src/Components/VideoPlayer/VideoPage.jsx
@@ -2,7 +2,7 @@ import { useEffect } from "react";
 import ReactPlayer from "react-player";
 import { useParams } from "react-router";
 import { useDataContext } from "../../Context/data-context";
-import { AddToPlaylist } from "../Playlist/addToPlaylist";
+import { AddToPlaylist } from "../Playlist/AddToPlaylist";
 import {
   formatDate,
   formatNumber,
